Refactored a bit; changed C program to match functional model.
1. Refactored the files so that C program verifications DO NOT
  depend (even indirectly) on importing the vcfloat proof system.

2. Edited the C program to better match the functional model.
 The fix was:  change  (force(x)+a) to (a+force(x)).
 With this change, the func.model-to-C refinement proof works
 without needing any difficult-to-prove lemmas, and WITHOUT
 needing a proof that x is finite (or that anything is finite,
 for that matter).

Some explanation:  When the functional model says (half*e)
for some expression e, and the C program says (e*half),
then we need to use Float32.mul_commut.  The premise of that
lemma is that _at least one_ of the arguments is finite.
For a constant value like "half", that's trivial to prove.

When the functional model says (a+F(x)) and the C program says
(F(x)+a), we need to use Float32.add_commut, whose premise
is that _at least one_ of the arguments is finite.  For
that we need a proof that (F(x)) is finite, for which we
need a proof that x is finite, and also a proof that  if x
is finite then (-1.0 * x) is finite; that latter lemma is true
but difficult to prove.

By matching the C program to the functional model (or vice versa)
we eliminate all commutations (except where one argument is
a constant), and so we no longer need to know that x is finite,
or that x*1=x, or that (-1)*x is finite.

That means, even if x becomes infinite or not-a-number, the C program
still matches the functional model.  You might think, "in that case
the functional model is not very meaningful", but that is not the
worry of the refinement proof.